backend/requirements.txt
fastapi
uvicorn
python-multipart
jinja2
pandas
openpyxl
reportlab
plotly
google-generativeai
itsdangerous
